+++ demisto_sdk/commands/create_id_set/tests/update_id_set_test.py
@@ -44,7 +44,9 @@ class TestIDSetCreator:
         id_set_creator.create_id_set()
         assert os.path.exists(self.file_path)
 
-    def test_create_id_set_no_output(self):
+    def test_create_id_set_no_output(self, mocker):
+        import demisto_sdk.commands.common.update_id_set as uis
+        mocker.patch.object(uis, 'cpu_count', return_value=1)
         id_set_creator = IDSetCreator()
 
         id_set = id_set_creator.create_id_set()
@@ -579,10 +581,8 @@ class TestLayouts:
         """
         Given
             - A layoutscontainer file called layoutscontainer-to-test.json
-
         When
             - parsing layoutscontainer files
-
         Then
             - parsing all the data from file successfully
         """