+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Added the *--skip-pack-dependencies* flag to **validate** command to skip pack dependencies validation. Relevant when using the *-g* flag.
 * Fixed an issue where **Validate** and **Format** commands failed integrations with `defaultvalue` field in fetch incidents related parameters.
 
 # 1.1.9